requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

pandas==2.2.2
openpyxl==3.1.5

pillow==10.4.0

google-cloud-storage==2.18.2

requests==2.32.3